added ina proxy option to use in conjunction with burp, I also refactored the code to make it module based.
