+++ src/main.rs
@@ -1,29 +1,15 @@
 use std::error::Error;
 use std::io::Write;
-use std::iter::Iterator;
 use std::process::exit;
 use std::time::Duration;
-
-use itertools;
-use differ::{Differ, Tag};
-
-use itertools::iproduct;
-use regex::Regex;
-
-use distance::sift3;
 
 use clap::App;
 use clap::Arg;
 
 use futures::stream::FuturesUnordered;
 use futures::StreamExt;
-use governor::Quota;
-use governor::RateLimiter;
-
-use reqwest::redirect;
 
 use tokio::fs::OpenOptions;
-use tokio::io::AsyncWriteExt;
 use tokio::sync::mpsc;
 
 use tokio::io::{AsyncBufReadExt, BufReader};
@@ -34,49 +20,16 @@ use tokio::{fs::File, task};
 use colored::Colorize;
 use indicatif::{ProgressBar, ProgressDrawTarget, ProgressStyle};
 
-// the Job struct which will be used to define our settings for the job
-#[derive(Clone, Debug)]
-struct JobSettings {
-    match_status: String,
-    drop_after_fail: String,
-}
+use crate::bruteforcer::BruteJob;
+use crate::bruteforcer::BruteResult;
+use crate::detector::Job;
+use crate::detector::JobResult;
 
-// the Job struct which will be used to send to the workers
-#[derive(Clone, Debug)]
-struct Job {
-    settings: Option<JobSettings>,
-    url: Option<String>,
-    payload: Option<String>,
-}
+mod bruteforcer;
+mod detector;
+mod utils;
 
-// the JobResult struct which contains the data to be saved to a file
-#[derive(Clone, Debug)]
-pub struct JobResult {
-    data: String,
-}
-
-// the BruteResult struct which contains the data to be saved to a file
-#[derive(Clone, Debug)]
-pub struct BruteResult {
-    data: String,
-}
-
-// the Job struct which will be used for directory bruteforcing
-#[derive(Clone, Debug)]
-struct BruteJob {
-    url: Option<String>,
-    word: Option<String>,
-}
-
-struct Threshold {
-    threshold_start: f32,
-    threshold_end: f32,
-}
-const CHANGE: Threshold = Threshold {
-    threshold_start: 50.0,
-    threshold_end: 50000.0,
-};
-
+// our fancy ascii banner to make it look hackery :D
 fn print_banner() {
     const BANNER: &str = r#"                             
                  __  __    __               __           
@@ -85,7 +38,7 @@ fn print_banner() {
   / /_/ / /_/ / /_/ / / / /_/ / /_/ (__  ) /_/  __/ /    
  / .___/\__,_/\__/_/ /_/_.___/\__,_/____/\__/\___/_/     
 /_/                                                          
-                     v0.4.1
+                     v0.4.2
                      ------
         path normalization pentesting tool                       
     "#;
@@ -119,6 +72,7 @@ fn print_banner() {
     );
 }
 
+// asynchronous entry point main where the magic happens.
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
     // print the banner
@@ -126,7 +80,7 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
 
     // parse the cli arguments
     let matches = App::new("pathbuster")
-        .version("0.4.1")
+        .version("0.4.2")
         .author("Blake Jacobs <krypt0mux@gmail.com>")
         .about("path-normalization pentesting tool")
         .arg(
@@ -175,6 +129,14 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
                 .takes_value(true)
                 .default_value("./wordlists/wordlist.txt")
                 .help("the file containing the wordlist used for directory bruteforcing"),
+        )
+        .arg(
+            Arg::with_name("proxy")
+                .short('p')
+                .long("proxy")
+                .default_value("1000")
+                .takes_value(true)
+                .help("The amount of concurrent requests"),
         )
         .arg(
             Arg::with_name("concurrency")
@@ -235,6 +197,14 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
                 "could not parse drop-after-fail, using default of 302,301"
             );
             "".to_string()
+        }
+    };
+
+    let http_proxy = match matches.get_one::<String>("proxy").map(|p| p.to_string()) {
+        Some(http_proxy) => http_proxy,
+        None => {
+            println!("{}", "could not parse http_proxy");
+            exit(1);
         }
     };
 
@@ -398,20 +368,22 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
     // spawn our workers
     let out_pb = pb.clone();
     let job_pb = pb.clone();
-    rt.spawn(
-        async move { send_url(job_tx, urls, payloads, rate, match_status, drop_after_fail).await },
-    );
+    rt.spawn(async move {
+        detector::send_url(job_tx, urls, payloads, rate, match_status, drop_after_fail).await
+    });
 
     // process the jobs
     let workers = FuturesUnordered::new();
 
     // process the jobs for scanning.
     for _ in 0..concurrency {
+        let http_proxy = http_proxy.clone();
         let jrx = job_rx.clone();
         let jtx: mpsc::Sender<JobResult> = result_tx.clone();
         let jpb = job_pb.clone();
         workers.push(task::spawn(async move {
-            run_tester(jpb, jrx, jtx, timeout).await
+            //  run the detector
+            detector::run_tester(jpb, jrx, jtx, timeout, http_proxy).await
         }));
     }
 
@@ -455,7 +427,7 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
                     exit(1);
                 }
             };
-            save_traversals(out_pb, outfile_handle_traversal, out_data).await;
+            detector::save_traversals(out_pb, outfile_handle_traversal, out_data).await;
         }
     }
     let outfile_path_brute = outfile_path_brute.clone();
@@ -478,18 +450,21 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
     let (brute_job_tx, brute_job_rx) = spmc::channel::<BruteJob>();
     let (brute_result_tx, brute_result_rx) = mpsc::channel::<BruteResult>(w);
     // start orchestrator tasks
-    rt.spawn(async move { send_word_to_url(brute_job_tx, results, brute_wordlist, rate).await });
-    rt.spawn(
-        async move { save_discoveries(out_pb, outfile_handle_brute, brute_result_rx).await },
-    );
+    rt.spawn(async move {
+        bruteforcer::send_word_to_url(brute_job_tx, results, brute_wordlist, rate).await
+    });
+    rt.spawn(async move {
+        bruteforcer::save_discoveries(out_pb, outfile_handle_brute, brute_result_rx).await
+    });
     // process the jobs for directory bruteforcing.
     let workers = FuturesUnordered::new();
     for _ in 0..concurrency {
+        let http_proxy = http_proxy.clone();
         let brx = brute_job_rx.clone();
         let btx: mpsc::Sender<BruteResult> = brute_result_tx.clone();
         let bpb = job_pb.clone();
         workers.push(task::spawn(async move {
-            run_bruteforcer(bpb, brx, btx, timeout).await
+            bruteforcer::run_bruteforcer(bpb, brx, btx, timeout, http_proxy).await
         }));
     }
     let worker_results: Vec<_> = workers.collect().await;
@@ -531,732 +506,4 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
     );
 
     Ok(())
-}
-
-// find the different characters between two strings
-fn str_change_percent(a: &str, b: &str) -> (bool, f32) {
-    let s = sift3(a, b);
-    if s > CHANGE.threshold_start && s < CHANGE.threshold_end {
-        return (true, s);
-    }
-    return (false, 0.0);
-}
-
-// this function will send the jobs to the workers
-async fn send_url(
-    mut tx: spmc::Sender<Job>,
-    urls: Vec<String>,
-    payloads: Vec<String>,
-    rate: u32,
-    match_status: String,
-    drop_after_fail: String,
-) -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
-    //set rate limit
-    let lim = RateLimiter::direct(Quota::per_second(std::num::NonZeroU32::new(rate).unwrap()));
-
-    // the job settings
-    let job_settings = JobSettings {
-        match_status: match_status.to_string(),
-        drop_after_fail: drop_after_fail,
-    };
-
-    // start the scan
-    for (url,payload) in iproduct!(urls, payloads)  {
-        let msg = Job {
-            settings: Some(job_settings.clone()),
-            url: Some(url.clone()),
-            payload: Some(payload.clone()),
-        };
-        if let Err(_) = tx.send(msg) {
-            continue;
-        }
-        lim.until_ready().await;
-    }
-    Ok(())
-}
-
-// this function will send the jobs to the workers
-async fn send_word_to_url(
-    mut tx: spmc::Sender<BruteJob>,
-    urls: Vec<String>,
-    wordlists: Vec<String>,
-    rate: u32,
-) -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
-    //set rate limit
-    let lim = RateLimiter::direct(Quota::per_second(std::num::NonZeroU32::new(rate).unwrap()));
-
-    // start the scan
-    for (word, url) in iproduct!(wordlists, urls) {
-        let url_cp = url.clone();
-        let msg = BruteJob {
-            url: Some(url_cp),
-            word: Some(word.clone()),
-        };
-        if let Err(_) = tx.send(msg) {
-            continue;
-        }
-        lim.until_ready().await;
-    }
-    Ok(())
-}
-
-// runs the directory bruteforcer on the job
-async fn run_bruteforcer(
-    pb: ProgressBar,
-    rx: spmc::Receiver<BruteJob>,
-    tx: mpsc::Sender<BruteResult>,
-    timeout: usize,
-) -> BruteResult {
-    let mut headers = reqwest::header::HeaderMap::new();
-    headers.insert(
-        reqwest::header::USER_AGENT,
-        reqwest::header::HeaderValue::from_static(
-            "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:95.0) Gecko/20100101 Firefox/95.0",
-        ),
-    );
-
-    //no certs
-    let client = reqwest::Client::builder()
-        .default_headers(headers)
-        .redirect(redirect::Policy::none())
-        .timeout(Duration::from_secs(timeout.try_into().unwrap()))
-        .danger_accept_invalid_hostnames(true)
-        .danger_accept_invalid_certs(true)
-        .build()
-        .unwrap();
-
-    while let Ok(job) = rx.recv() {
-        let job_url = job.url.unwrap();
-        let job_word = job.word.unwrap();
-        let job_url_new = job_url.clone();
-
-        let mut web_root_url: String = String::from("");
-        let mut internal_web_root_url: String = String::from(job_url);
-        let url = match reqwest::Url::parse(&job_url_new) {
-            Ok(url) => url,
-            Err(_) => {
-                continue;
-            }
-        };
-
-        pb.inc(1);
-
-        let schema = url.scheme().to_string();
-        let host = match url.host_str() {
-            Some(host) => host,
-            None => continue,
-        };
-
-        web_root_url.push_str(&schema);
-        web_root_url.push_str("://");
-        web_root_url.push_str(&host);
-        web_root_url.push_str("/");
-        web_root_url.push_str(&job_word);
-
-        internal_web_root_url.push_str(&job_word);
-        let internal_url = internal_web_root_url.clone();
-        let internal_web_url = internal_url.clone();
-
-        pb.set_message(format!(
-            "{} {}",
-            "directory bruteforcing ::".bold().white(),
-            internal_url.bold().blue(),
-        ));
-
-        let get = client.get(internal_web_url);
-        let internal_get = client.get(internal_web_root_url);
-        let public_get = client.get(web_root_url);
-
-        let public_req = match public_get.build() {
-            Ok(req) => req,
-            Err(_) => {
-                continue;
-            }
-        };
-
-        let internal_req = match internal_get.build() {
-            Ok(req) => req,
-            Err(_) => {
-                continue;
-            }
-        };
-
-        let public_resp = match client.execute(public_req).await {
-            Ok(public_resp) => public_resp,
-            Err(_) => {
-                continue;
-            }
-        };
-
-        let internal_resp = match client.execute(internal_req).await {
-            Ok(internal_resp) => internal_resp,
-            Err(_) => {
-                continue;
-            }
-        };
-
-        let public_resp_text = match public_resp.text().await {
-            Ok(public_cl) => public_cl,
-            Err(_) => continue,
-        };
-
-        let internal_resp_text = match internal_resp.text().await {
-            Ok(internal_cl) => internal_cl,
-            Err(_) => continue,
-        };
-
-        let req = match get.build() {
-            Ok(req) => req,
-            Err(_) => {
-                continue;
-            }
-        };
-
-        let resp = match client.execute(req).await {
-            Ok(resp) => resp,
-            Err(_) => {
-                continue;
-            }
-        };
-
-        let (ok, distance_between_responses) =
-            str_change_percent(&internal_resp_text, &public_resp_text);
-        if ok && resp.status().as_str() != "404" && resp.status().as_str() != "400" {
-            let internal_resp_text_lines = internal_resp_text.lines().collect::<Vec<_>>();
-            let public_resp_text_lines = public_resp_text.lines().collect::<Vec<_>>();
-            let character_differences =
-                Differ::new(&internal_resp_text_lines, &public_resp_text_lines);
-            pb.println(format!(
-                "\n{}{}{} {}",
-                "(".bold().white(),
-                "*".bold().blue(),
-                ")".bold().white(),
-                "found some response changes:".bold().green(),
-            ));
-            for span in character_differences.spans() {
-                match span.tag {
-                    Tag::Equal => (),  // ignore
-                    Tag::Insert => (), // ignore
-                    Tag::Delete => (), // ignore
-                    Tag::Replace => {
-                        let mut i = 0;
-                        for line in &internal_resp_text_lines[span.b_start..span.b_end] {
-                            i = i + 1;
-                            if i < 100 {
-                                if line.to_string() == "" {
-                                    pb.println(format!("\n{}", line.bold().white(),));
-                                } else {
-                                    pb.println(format!("{}", line.bold().white(),));
-                                }
-                            } else {
-                                pb.println(format!("\n{}\n", "...".bold().white(),));
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-            pb.println(format!("\n"));
-            pb.println(format!(
-                "{} {}{}{} {} {}",
-                "found something interesting".bold().green(),
-                "(".bold().white(),
-                distance_between_responses.to_string().bold().white(),
-                ")".bold().white(),
-                "deviations from webroot ::".bold().white(),
-                internal_url.bold().blue(),
-            ));
-            pb.inc_length(1);
-
-            // send the result message through the channel to the workers.
-            let result_msg = BruteResult {
-                data: internal_url.to_owned(),
-            };
-            let result = result_msg.clone();
-            if let Err(_) = tx.send(result_msg).await {
-                continue;
-            }
-
-            return result;
-        }
-    }
-    return BruteResult {
-        data: "".to_string(),
-    };
-}
-
-// this function will test for path normalization vulnerabilities
-async fn run_tester(
-    pb: ProgressBar,
-    rx: spmc::Receiver<Job>,
-    tx: mpsc::Sender<JobResult>,
-    timeout: usize,
-) -> JobResult {
-    let mut headers = reqwest::header::HeaderMap::new();
-    headers.insert(
-        reqwest::header::USER_AGENT,
-        reqwest::header::HeaderValue::from_static(
-            "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:95.0) Gecko/20100101 Firefox/95.0",
-        ),
-    );
-
-    //no certs
-    let client = reqwest::Client::builder()
-        .default_headers(headers)
-        .redirect(redirect::Policy::none())
-        .timeout(Duration::from_secs(timeout.try_into().unwrap()))
-        .danger_accept_invalid_hostnames(true)
-        .danger_accept_invalid_certs(true)
-        .build()
-        .unwrap();
-
-    while let Ok(job) = rx.recv() {
-        let job_url = job.url.unwrap();
-        let job_payload = job.payload.unwrap();
-        let job_settings = job.settings.unwrap();
-        let job_url_new = job_url.clone();
-        let job_payload_new = job_payload.clone();
-        pb.inc(1);
-        let url = match reqwest::Url::parse(&job_url_new) {
-            Ok(url) => url,
-            Err(_) => {
-                continue;
-            }
-        };
-        let mut job_url_with_path: String = String::from("");
-        let mut job_url_without_path: String = String::from("");
-        let schema = url.scheme().to_string();
-        let path = url.path().to_string();
-        let host = match url.host_str() {
-            Some(host) => host,
-            None => continue,
-        };
-
-        job_url_with_path.push_str(&schema);
-        job_url_with_path.push_str("://");
-        job_url_with_path.push_str(&host);
-        job_url_with_path.push_str(&path);
-        job_url_without_path.push_str(&schema);
-        job_url_without_path.push_str("://");
-        job_url_without_path.push_str(&host);
-        job_url_without_path.push_str("/");
-
-        let path_cnt = path.split("/").count() + 5;
-        let mut payload = String::from(job_payload);
-        let new_url = String::from(&job_url);
-        let mut track_status_codes = 0;
-        for _ in 0..path_cnt {
-            let job_url_without_path = job_url_without_path.clone();
-            let mut new_url = new_url.clone();
-            if !new_url.as_str().ends_with("/") {
-                new_url.push_str("/");
-            }
-            new_url.push_str(&payload);
-            pb.set_message(format!(
-                "{} {}",
-                "scanning ::".bold().white(),
-                new_url.bold().blue(),
-            ));
-            let new_url2 = new_url.clone();
-            let get = client.get(new_url);
-            let req = match get.build() {
-                Ok(req) => req,
-                Err(_) => {
-                    continue;
-                }
-            };
-            let resp = match client.execute(req).await {
-                Ok(resp) => resp,
-                Err(_) => {
-                    continue;
-                }
-            };
-            let pub_get = client.get(job_url_without_path);
-            let pub_req = match pub_get.build() {
-                Ok(pub_req) => pub_req,
-                Err(_) => {
-                    continue;
-                }
-            };
-            let pub_resp = match client.execute(pub_req).await {
-                Ok(pub_resp) => pub_resp,
-                Err(_) => {
-                    continue;
-                }
-            };
-
-            let content_length = match resp.content_length() {
-                Some(content_length) => content_length.to_string(),
-                None => { "" }.to_owned(),
-            };
-            let backonemore_url = new_url2.clone();
-
-            if job_settings.match_status.contains(resp.status().as_str()) {
-                // strip the suffix hax and traverse back one more level
-                // to reach the internal doc root.
-                let backonemore = match backonemore_url.strip_suffix(job_payload_new.as_str()) {
-                    Some(backonemore) => backonemore,
-                    None => "",
-                };
-                let get = client.get(backonemore);
-                let request = match get.build() {
-                    Ok(request) => request,
-                    Err(_) => {
-                        continue;
-                    }
-                };
-                let response_title = match client.execute(request).await {
-                    Ok(response_title) => response_title,
-                    Err(_) => {
-                        continue;
-                    }
-                };
-
-                let result_url = backonemore.clone();
-                let get = client.get(backonemore);
-                let request = match get.build() {
-                    Ok(request) => request,
-                    Err(_) => {
-                        continue;
-                    }
-                };
-                let response = match client.execute(request).await {
-                    Ok(response) => response,
-                    Err(_) => {
-                        continue;
-                    }
-                };
-
-                let internal_get = client.get(backonemore);
-
-                let internal_req = match internal_get.build() {
-                    Ok(internal_req) => internal_req,
-                    Err(_) => {
-                        continue;
-                    }
-                };
-                let internal_resp = match client.execute(internal_req).await {
-                    Ok(internal_resp) => internal_resp,
-                    Err(_) => {
-                        continue;
-                    }
-                };
-
-                let internal_cl = match internal_resp.text().await {
-                    Ok(internal_cl) => internal_cl,
-                    Err(_) => continue,
-                };
-
-                let public_cl = match pub_resp.text().await {
-                    Ok(public_cl) => public_cl,
-                    Err(_) => continue,
-                };
-
-                // we hit the internal doc root.
-                let (ok, distance_between_responses) = str_change_percent(&internal_cl, &public_cl);
-                if response.status().as_str() != "400"
-                    && ok
-                    && result_url.contains(&job_payload_new)
-                {
-                    // track the status codes
-                    if job_settings.drop_after_fail == response.status().as_str() {
-                        track_status_codes += 1;
-                        if track_status_codes >= 5 {
-                            return JobResult {
-                                data: "".to_string(),
-                            };
-                        }
-                    }
-                    pb.println(format!(
-                        "{} {}",
-                        "found internal doc root :: ".bold().green(),
-                        result_url.bold().blue(),
-                    ));
-                    let mut title = String::from("");
-                    let content = match response_title.text().await {
-                        Ok(content) => content,
-                        Err(_) => "".to_string(),
-                    };
-                    let re = Regex::new(r"<title>(.*?)</title>").unwrap();
-                    for cap in re.captures_iter(&content) {
-                        title.push_str(&cap[1]);
-                    }
-                    // fetch the server from the headers
-                    let server = match response.headers().get("Server") {
-                        Some(server) => match server.to_str() {
-                            Ok(server) => server,
-                            Err(_) => "Unknown",
-                        },
-                        None => "Unknown",
-                    };
-
-                    let internal_resp_text_lines = internal_cl.lines().collect::<Vec<_>>();
-                    let public_resp_text_lines = public_cl.lines().collect::<Vec<_>>();
-                    let character_differences =
-                        Differ::new(&public_resp_text_lines, &internal_resp_text_lines);
-                    pb.println(format!(
-                        "\n{}{}{} {}",
-                        "(".bold().white(),
-                        "*".bold().blue(),
-                        ")".bold().white(),
-                        "found some response changes:".bold().green(),
-                    ));
-                    for span in character_differences.spans() {
-                        match span.tag {
-                            Tag::Equal => (),  // ignore
-                            Tag::Insert => (), // ignore
-                            Tag::Delete => (), // ignore
-                            Tag::Replace => {
-                                let mut i = 0;
-                                for line in &internal_resp_text_lines[span.b_start..span.b_end] {
-                                    i = i + 1;
-                                    if i < 100 {
-                                        if line.to_string() == "" {
-                                            pb.println(format!("\n{}", line.bold().white(),));
-                                        } else {
-                                            pb.println(format!("{}", line.bold().white(),));
-                                        }
-                                    } else {
-                                        pb.println(format!("\n{}", "...".bold().white(),));
-                                        break;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    pb.println(format!("\n"));
-                    if response.status().is_client_error() {
-                        pb.println(format!(
-                            "{}{}{} {}{}{}\n{}{}{} {}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t",
-                            "[".bold().white(),
-                            "OK".bold().green(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            result_url.bold().cyan(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            "*".bold().green(),
-                            "]".bold().white(),
-                            "Response:".bold().white(),
-                            "payload:".bold().white(),
-                            "[".bold().white(),
-                            job_payload_new.bold().blue(),
-                            "]".bold().white(),
-                            "status:".bold().white(),
-                            "[".bold().white(),
-                            response.status().as_str().bold().blue(),
-                            "]".bold().white(),
-                            "content_length:".bold().white(),
-                            "[".bold().white(),
-                            content_length.yellow(),
-                            "]".bold().white(),
-                            "server:".bold().white(),
-                            "[".bold().white(),
-                            server.bold().purple(),
-                            "]".bold().white(),
-                            "title:".bold().white(),
-                            "[".bold().white(),
-                            title.bold().purple(),
-                            "]".bold().white(),
-                            "deviation:".bold().white(),
-                            "[".bold().white(),
-                            distance_between_responses.to_string().bold().purple(),
-                            "]".bold().white(),
-                        ));
-                    }
-                    if response.status().is_success() {
-                        pb.println(format!(
-                            "{}{}{} {}{}{}\n{}{}{} {}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t",
-                            "[".bold().white(),
-                            "OK".bold().green(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            result_url.bold().cyan(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            "*".bold().green(),
-                            "]".bold().white(),
-                            "Response:".bold().white(),
-                            "payload:".bold().white(),
-                            "[".bold().white(),
-                            job_payload_new.bold().blue(),
-                            "]".bold().white(),
-                            "status:".bold().white(),
-                            "[".bold().white(),
-                            response.status().as_str().bold().green(),
-                            "]".bold().white(),
-                            "content_length:".bold().white(),
-                            "[".bold().white(),
-                            content_length.yellow(),
-                            "]".bold().white(),
-                            "server:".bold().white(),
-                            "[".bold().white(),
-                            server.bold().purple(),
-                            "]".bold().white(),
-                            "title:".bold().white(),
-                            "[".bold().white(),
-                            title.bold().purple(),
-                            "]".bold().white(),
-                            "deviation:".bold().white(),
-                            "[".bold().white(),
-                            distance_between_responses.to_string().bold().purple(),
-                            "]".bold().white(),
-                        ));
-                    }
-                    if response.status().is_redirection() {
-                        pb.println(format!(
-                            "{}{}{} {}{}{}\n{}{}{} {}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t",
-                            "[".bold().white(),
-                            "OK".bold().green(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            result_url.bold().cyan(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            "*".bold().green(),
-                            "]".bold().white(),
-                            "Response:".bold().white(),
-                            "payload:".bold().white(),
-                            "[".bold().white(),
-                            job_payload_new.bold().blue(),
-                            "]".bold().white(),
-                            "status:".bold().white(),
-                            "[".bold().white(),
-                            response.status().as_str().bold().blue(),
-                            "]".bold().white(),
-                            "content_length:".bold().white(),
-                            "[".bold().white(),
-                            content_length.yellow(),
-                            "]".bold().white(),
-                            "server:".bold().white(),
-                            "[".bold().white(),
-                            server.bold().purple(),
-                            "]".bold().white(),
-                            "title:".bold().white(),
-                            "[".bold().white(),
-                            title.bold().purple(),
-                            "]".bold().white(),
-                            "deviation:".bold().white(),
-                            "[".bold().white(),
-                            distance_between_responses.to_string().bold().purple(),
-                            "]".bold().white(),
-                        ));
-                    }
-                    if response.status().is_server_error() {
-                        pb.println(format!(
-                            "{}{}{} {}{}{}\n{}{}{} {}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t",
-                            "[".bold().white(),
-                            "OK".bold().green(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            result_url.bold().cyan(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            "*".bold().green(),
-                            "]".bold().white(),
-                            "Response:".bold().white(),
-                            "payload:".bold().white(),
-                            "[".bold().white(),
-                            job_payload_new.bold().blue(),
-                            "]".bold().white(),
-                            "status:".bold().white(),
-                            "[".bold().white(),
-                            response.status().as_str().bold().red(),
-                            "]".bold().white(),
-                            "content_length:".bold().white(),
-                            "[".bold().white(),
-                            content_length.yellow(),
-                            "]".bold().white(),
-                            "server:".bold().white(),
-                            "[".bold().white(),
-                            server.bold().purple(),
-                            "]".bold().white(),
-                            "title:".bold().white(),
-                            "[".bold().white(),
-                            title.bold().purple(),
-                            "]".bold().white(),
-                            "deviation:".bold().white(),
-                            "[".bold().white(),
-                            distance_between_responses.to_string().bold().purple(),
-                            "]".bold().white(),
-                        ));
-                    }
-                    if response.status().is_informational() {
-                        pb.println(format!(
-                            "{}{}{} {}{}{}\n{}{}{} {}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t {} {}{}{}\n\t",
-                            "[".bold().white(),
-                            "OK".bold().green(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            result_url.bold().cyan(),
-                            "]".bold().white(),
-                            "[".bold().white(),
-                            "*".bold().green(),
-                            "]".bold().white(),
-                            "Response:".bold().white(),
-                            "payload:".bold().white(),
-                            "[".bold().white(),
-                            job_payload_new.bold().blue(),
-                            "]".bold().white(),
-                            "status:".bold().white(),
-                            "[".bold().white(),
-                            response.status().as_str().bold().purple(),
-                            "]".bold().white(),
-                            "content_length:".bold().white(),
-                            "[".bold().white(),
-                            content_length.yellow(),
-                            "]".bold().white(),
-                            "server:".bold().white(),
-                            "[".bold().white(),
-                            server.bold().purple(),
-                            "]".bold().white(),
-                            "title:".bold().white(),
-                            "[".bold().white(),
-                            title.bold().purple(),
-                            "]".bold().white(),
-                            "deviation:".bold().white(),
-                            "[".bold().white(),
-                            distance_between_responses.to_string().bold().purple(),
-                            "]".bold().white(),
-                        ));
-                    }
-                    // send the result message through the channel to the workers.
-                    let result_msg = JobResult {
-                        data: result_url.to_owned(),
-                    };
-                    let result_job = result_msg.clone();
-                    if let Err(_) = tx.send(result_msg).await {
-                        continue;
-                    }
-                    pb.inc_length(1);
-                    return result_job;
-                }
-            }
-            payload.push_str(&job_payload_new);
-        }
-    }
-    return JobResult {
-        data: "".to_string(),
-    };
-}
-
-async fn save_traversals(_: ProgressBar, mut outfile: File, traversal: String) {
-    let mut outbuf = traversal.as_bytes().to_owned();
-    outbuf.extend_from_slice(b"\n");
-    if let Err(_) = outfile.write(&outbuf).await {
-        // pb.println(format!("failed to write output '{:?}': {:?}", outbuf, e));
-        return;
-    }
-}
-
-// Saves the output to a file
-async fn save_discoveries(_: ProgressBar, mut outfile: File, mut brx: mpsc::Receiver<BruteResult>) {
-    while let Some(result) = brx.recv().await {
-        let mut outbuf = result.data.as_bytes().to_owned();
-        outbuf.extend_from_slice(b"\n");
-        if let Err(_) = outfile.write(&outbuf).await {
-            // pb.println(format!("failed to write output '{:?}': {:?}", outbuf, e));
-            continue;
-        }
-    }
 }
